# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Research Assistant API",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/research": {
      "get": {
        "summary": "List Research",
        "operationId": "list_research_research_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobSummary"
                  },
                  "type": "array",
                  "title": "Response List Research Research Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Research",
        "operationId": "create_research_research_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research/{job_id}": {
      "get": {
        "summary": "Get Research",
        "operationId": "get_research_research__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research/{job_id}/events": {
      "get": {
        "summary": "Get Research Events",
        "operationId": "get_research_events_research__job_id__events_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobEvent"
                  },
                  "title": "Response Get Research Events Research  Job Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research/{job_id}/stream": {
      "get": {
        "summary": "Stream Research",
        "operationId": "stream_research_research__job_id__stream_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EnqueueResponse": {
        "properties": {
          "jobId": {
            "type": "string",
            "title": "Jobid"
          }
        },
        "type": "object",
        "required": [
          "jobId"
        ],
        "title": "EnqueueResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobDetail": {
        "properties": {
          "jobId": {
            "type": "string",
            "title": "Jobid"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "done",
              "failed"
            ],
            "title": "Status"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchResultPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "query",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "title": "JobDetail"
      },
      "JobEvent": {
        "properties": {
          "event": {
            "type": "string",
            "title": "Event"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "event",
          "timestamp"
        ],
        "title": "JobEvent"
      },
      "JobSummary": {
        "properties": {
          "jobId": {
            "type": "string",
            "title": "Jobid"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "done",
              "failed"
            ],
            "title": "Status"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "query",
          "status",
          "createdAt"
        ],
        "title": "JobSummary"
      },
      "ResearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query",
            "description": "The research question"
          },
          "notifyEmail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notifyemail",
            "description": "Optional email address. When set and AgentMail is configured, the finished research report is emailed to this address."
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "ResearchRequest"
      },
      "ResearchResultPayload": {
        "properties": {
          "markdown": {
            "type": "string",
            "title": "Markdown"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/ResearchSource"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "markdown"
        ],
        "title": "ResearchResultPayload"
      },
      "ResearchSource": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "title",
          "url"
        ],
        "title": "ResearchSource"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}